fix scripts(reflect pr comment)
