+++ scripts/releaseZip.js
@@ -5,23 +5,14 @@ const fs = require("fs");
 const os = require("os");
 const path = require("path");
 const sh = require("shelljs");
-const archiver = require("archiver");
+const execCommand = require("./util/execCommand");
+const compressToZip = require("./util/compressToZip");
 
 const tmpDirPath = fs.mkdtempSync(path.join(os.tmpdir(), "tkoolmv-namagame-kit"));
 const tkoolmvConverDirPath = path.resolve(__dirname, "..", "module", "tkoolmv-namagame-converter");
 const packageJson = require(path.resolve(__dirname, "..", "package.json"));
 
-function execCommand(command) {
-	const result = sh.exec(command);
-	if (result.code !== 0) {
-		console.error(result.stderr);
-		process.exit(1);
-	}
-	return result.stdout.trim();
-}
-
 (async() => {
-	// zipファイルの作成
 	const version = packageJson["version"];
 	const zipDirPath = path.join(tmpDirPath, `tkoolmv-namagame-kit-${version}`);
 	if (fs.existsSync(zipDirPath)) {
@@ -33,6 +24,9 @@ function execCommand(command) {
 	execCommand("npm i @akashic/tkoolmv-namagame-runtime@latest"); // コアスクリプトを runtime モジュールから取得するため npm install する
 	const tkoolmvRuntimeDirPath = path.resolve(__dirname, "../node_modules/@akashic/tkoolmv-namagame-runtime/dist/tkoolmv-namagame-runtime");
 	sh.cp("-R", path.join(tkoolmvRuntimeDirPath, "*"), zipDirPath);
+	// ReleaseNoteの作成の際にランタイムのバージョン情報が必要なので予めここで取得しておく
+	const runtimePackageJson = require(path.resolve(__dirname, "../node_modules/@akashic/tkoolmv-namagame-runtime/package.json"));
+	const runtimeVersion = runtimePackageJson["version"];
 	execCommand("npm uninstall @akashic/tkoolmv-namagame-runtime"); // 以降で runtime は不要なので uninstall
 
 	// コンバーターの同梱
@@ -45,24 +39,14 @@ function execCommand(command) {
 
 	// zip圧縮処理
 	const zipPath = path.join(tmpDirPath, `tkoolmv-namagame-kit.zip`);
-	const ostream = fs.createWriteStream(zipPath);
-	// zip圧縮完了の通知が来るまで待機するための処理
-	const streamClosePromise = new Promise(resolve => ostream.on("close", resolve));
-	const archive = archiver("zip");
-	await archive.pipe(ostream);
-	await archive.glob(`${path.basename(zipDirPath)}/**`, {cwd: path.relative(process.cwd(), path.dirname(zipDirPath))});
-	await archive.finalize();
-	sh.rm("-Rf", zipDirPath);
-	// zipが不完全な状態でアップロードされるのを防ぐために、zip圧縮完了の通知が来るまで待機
-	await streamClosePromise;
-	console.log(`Completed: ${zipPath}`);
+	await compressToZip(zipDirPath, zipPath);
+	sh.rm("-Rf", zipDirPath); // zip圧縮後zipDirPathは不要なので削除
 
 	// ReleaseNoteの作成とzipファイルのアップロード
-	const runtimeVersion = execCommand("npm info @akashic/tkoolmv-namagame-runtime@latest version");
 	const releaseNoteContent = `* @akashic/tkoolmv-namagame-kit@${version}
   * [@akashic/tkoolmv-namagame-converter@${converterVersion}](https://github.com/akashic-games/tkoolmv-namagame-converter/releases/tag/v${converterVersion})
   * [@akashic/tkoolmv-namagame-runtime@${runtimeVersion}](https://github.com/akashic-games/tkoolmv-namagame-runtime/releases/tag/v${runtimeVersion})
-`;
+`
 	execCommand(`echo ${process.env.GITHUB_CLI_TOKEN} | gh auth login --with-token -h github.com`);
 	sh.exec(`gh release create "v${version}" -t "Release v${version}" --target "main" -F "${releaseNoteContent}"`);
 	execCommand(`gh release upload "v${version}" "${zipPath}"`);
